メインファイルをsrc/index.jsに変更
