--- main.ts
@@ -1,10 +1,0 @@
-const express = require("express");
-const app = express();
-
-app.get("/", function (req: any, res: any) {
-  res.send("Hello World!");
-});
-
-app.listen(3000, function () {
-  console.log("Example app listening on port 3000!");
-});
