+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node dist/main.js",
+    "start": "node dist/src/index.js",
     "build": "npx tsc",
     "dev": "npx tsc-watch --onSuccess \"npm start\""
   },
